python_script/requirements.txt
opencv-python
numpy
deepface
mediapipe
python-socketio